type error fix

+++ mros2-mbed/workspace/proj_node/Node_MotorControl/app.cpp
@@ -22,7 +22,7 @@
 #include <string>
 
 void splitData(std::string cmData);
-void frontControl(std::string frontDirection, float diff_degree);
+void frontControl(std::string frontDirection, uint8_t diff_degree);
 void motorControl(int period_PWM, float dutycycle_PWM, std::string backDirection);
 void parseCommandData(const std::string& cmData, const char*& outFrontDir, uint8_t& outFrontAng, uint8_t& outDutyCycle, const char*& outBackDir);
